Separate balance of really owned and with further unlock conditions restricted outputs
