+++ src/account/types/mod.rs
@@ -29,19 +29,30 @@ use std::{collections::HashMap, str::FromStr};
 #[derive(Default, Clone, Debug, Serialize, Deserialize)]
 pub struct AccountBalance {
     // Total amount
-    pub(crate) total: u64,
-    // balance that can currently spend
-    pub(crate) available: u64,
+    pub total: u64,
+    // From additional unlock conditions restricted amount
+    #[serde(rename = "restrictedAmount")]
+    pub restricted_amount: u64,
+    // balance that can currently be spend
+    pub available: u64,
+    // From additional unlock conditions restricted current required storage deposit amount
+    #[serde(rename = "requiredStorageDeposit")]
+    pub required_storage_deposit: u64,
     // currently required storage deposit amount
-    pub(crate) required_storage_deposit: u64,
+    #[serde(rename = "restrictedRequiredStorageDeposit")]
+    pub restricted_required_storage_deposit: u64,
     // Native tokens
-    pub(crate) native_tokens: HashMap<TokenId, U256>,
+    #[serde(rename = "nativeTokens")]
+    pub native_tokens: HashMap<TokenId, U256>,
+    // From additional unlock conditions native tokens
+    #[serde(rename = "restrictedNativeTokens")]
+    pub restricted_native_tokens: HashMap<TokenId, U256>,
     // Output ids of owned nfts // would the nft id/address be better?
-    pub(crate) nfts: Vec<OutputId>,
+    pub nfts: Vec<OutputId>,
     // Output ids of alias nfts // would the alias id/address be better?
-    pub(crate) aliases: Vec<OutputId>,
+    pub aliases: Vec<OutputId>,
     // Foundry outputs
-    pub(crate) foundries: Vec<OutputId>,
+    pub foundries: Vec<OutputId>,
 }
 
 /// An output with metadata
